// web/src/app/shared/api-client/model/exercise-input-feedback-text-dto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ExerciseInputFeedbackTextDto { 
    correctValues: Array<string>;
    incorrectValues: Array<string>;
    isRetypeAnswerEnabled?: boolean;
}