// web/src/app/shared/api-client/model/exercise-input-write-text-dto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ExerciseInputFeedbackTextDto } from './exercise-input-feedback-text-dto';
import { ExerciseInputDto } from './exercise-input-dto';


export interface ExerciseInputWriteTextDto extends ExerciseInputDto { 
    text: string;
    feedback?: ExerciseInputFeedbackTextDto;
}